Move useSandbox to test-helpers package
**Why**: For better organization, and so that it's accessible to package specs.
