+++ app/javascript/packages/test-helpers/index.ts
@@ -1,2 +1,3 @@
 export { default as useDefineProperty } from './use-define-property';
 export { default as usePropertyValue } from './use-property-value';
+export { default as useSandbox } from './use-sandbox';
