+++ app/javascript/packages/test-helpers/use-sandbox.ts
@@ -1,0 +1,32 @@
+import sinon from 'sinon';
+import type { SinonSandboxConfig } from 'sinon';
+
+/**
+ * Returns an instance of a Sinon sandbox, and automatically restores all stubbed methods after each
+ * test case.
+ */
+function useSandbox(config?: SinonSandboxConfig) {
+  const { useFakeTimers = false, ...remainingConfig } = config ?? {};
+  const sandbox = sinon.createSandbox(remainingConfig);
+
+  beforeEach(() => {
+    // useFakeTimers overrides global timer functions as soon as sandbox is created, thus leaking
+    // across tests. Instead, wait until tests start to initialize.
+    if (useFakeTimers) {
+      sandbox.useFakeTimers();
+    }
+  });
+
+  afterEach(() => {
+    sandbox.reset();
+    sandbox.restore();
+
+    if (useFakeTimers) {
+      sandbox.clock.restore();
+    }
+  });
+
+  return sandbox;
+}
+
+export default useSandbox;
