+++ spec/javascripts/packs/form-steps-wait-spec.js
@@ -1,6 +1,5 @@
 import { fireEvent, findByRole } from '@testing-library/dom';
-import { useDefineProperty } from '@18f/identity-test-helpers';
-import { useSandbox } from '../support/sinon';
+import { useDefineProperty, useSandbox } from '@18f/identity-test-helpers';
 import {
   FormStepsWait,
   getDOMFromHTML,
