+++ spec/javascripts/support/sinon.js
@@ -1,35 +1,3 @@
-import sinon from 'sinon';
-
-/**
- * Returns an instance of a Sinon sandbox, and automatically restores all stubbed methods after each
- * test case.
- *
- * @param {sinon.SinonSandboxConfig=} config
- */
-export function useSandbox(config) {
-  const { useFakeTimers = false, ...remainingConfig } = config ?? {};
-  const sandbox = sinon.createSandbox(remainingConfig);
-
-  beforeEach(() => {
-    // useFakeTimers overrides global timer functions as soon as sandbox is created, thus leaking
-    // across tests. Instead, wait until tests start to initialize.
-    if (useFakeTimers) {
-      sandbox.useFakeTimers();
-    }
-  });
-
-  afterEach(() => {
-    sandbox.reset();
-    sandbox.restore();
-
-    if (useFakeTimers) {
-      sandbox.clock.restore();
-    }
-  });
-
-  return sandbox;
-}
-
 /**
  * Chai plugin which allows a combination of `calledWith` and `eventually` to expect an eventual
  * spy (stub) call.
